third_party/blink/web_tests/html/selectlist/selectlist-slot-warning-button-expected.txt

CONSOLE WARNING: A <selectlist>'s default button was removed and a new one was not provided. This <selectlist> will not be fully functional.
This is a testharness.js-based test.
Harness: the test ran to completion.
